Re-license the library as LGPLv2-or-later
This facilitates the use from other projects such as ASL licensed
software.  We think there is no point in others having to re-implement
the GTP specific message parsing/encoding on top of libmnl, which is
LGPL anyway.

Please note that the tools are still licensed under GPL, just the
library has been changed.