configure.ac: Depend on osmo-release.mk
The osmo-release.mk script can be found in libosmocore and it is used in
most osmocom related projects to help create new releases.

LIBVERSION is moved to src/Makefile.am since osmo-release.mk greps for
Makefile.am to find LIBVERSIOn changes. In any case, we don't need in
the global include makefile.

Change-Id: I5d163535c34369c74ece26574c8b1601bc1697b7